fix(parity): three Pine-semantics gaps found by the 2026-06-10 TV-export audit — PASS at 0 flips
1. ta.pivothigh/low tie rule: >= left, > right (later twin wins). New
   pivot_high_pine/pivot_low_pine used on the DETECTION path (labels keep
   the loose rule); proven 346/346 pivots, drift mismatches 2474 -> 0.
2. precompute_matrices is now Pine-faithful: ta.cum cumsum-difference SMA
   (first valid at bar s), partial-window pir min/max, hi!=lo?:0.5, float64.
3. ta.atr is Wilder RMA, not SMA-of-TR (flipped the LOW volatility vote);
   ta.stdev is population (ddof=0).

Audit vs SP_SPX 1D TV export (25,249 bars, v17-GPU Run 1 preset):
signal flips 0/0, counts 71/71 and 47/47 -> PARITY PASS.
Golden baseline recaptured; GPU layer follows the f64 oracle matrix;
full suite 146 passed.

+++ tests/test_v17_gpu_eval_phase1.py
@@ -214,7 +214,7 @@ def test_tf32_and_low_precision_reductions_disabled() -> None:
 
 def test_feature_dtype_split(built) -> None:
     for name, (_, _, _, tp) in built.items():
-        assert tp.feat.pir_matrix.dtype == np.float32, name      # P1: PIR f32
+        assert tp.feat.pir_matrix.dtype == np.float64, name      # P1: PIR f64 (Pine-faithful)
         for _, feat_attr, kind in _FEATURE_ATTRS:
             arr = getattr(tp.feat, feat_attr)
             assert arr.dtype == np.dtype(kind), (name, feat_attr)
@@ -231,7 +231,7 @@ def test_pir_matrix_torch_byte_identical_on_three_slices(built) -> None:
     for name, (df, art, _, tp) in built.items():
         pir = build_pir_matrix_torch(
             df["close"], art.scales_list[0], art.scales_list[-1])
-        assert pir.dtype == np.float32, name
+        assert pir.dtype == np.float64, name
         assert np.array_equal(pir, art.pir_matrix, equal_nan=True), name
         # the evaluator's resident matrix is byte-identical too
         assert np.array_equal(tp.feat.pir_matrix, art.pir_matrix, equal_nan=True), name
